+++ src/aliases/navigation/go_to_id_or_area.lua
@@ -11,7 +11,7 @@ end
 
 local where = matches[2]:lower()
 
-if mapper.settings.debug then
+if mapper.debugging() then
 	mapper.gotoPerf = mapper.gotoPerf or createStopWatch()
 	startStopWatch(mapper.gotoPerf)
 end
@@ -32,6 +32,6 @@ else
 		end
 	end
 end
-if mapper.settings.debug then
+if mapper.debugging() then
 	mapper.echo("goto alias took " .. stopStopWatch(mapper.gotoPerf) .. "s to run.")
 end